feat: fully functional node 2. exercise 8 complete

+++ Node2/motor_controller.h
@@ -12,7 +12,7 @@
 
 static uint32_t last_time;
 static double integral;
-static double ref;
+static long double ref;
 
 void motor_init();
